Init ts
