+++ .gitignore
@@ -14,3 +14,6 @@ public/upload/uploads/
 includes/size.json
 .env
 dump.rdb
+
+node_modules/
+out/
