+++ includes/menu.php
@@ -64,5 +64,4 @@
   const root = "<?php echo ROOT ?>";
 </script>
 
-<script src="<?php echo ROOT ?>/js/formatter.js" defer></script>
 <script src="<?php echo ROOT ?>/js/menu.js"></script>
